for some reason, the __sync_lock_test_and_set gcc intrinsics is not available for uint8_t, while it's working on uint32_t.

+++ common/rwlock.c
@@ -6,7 +6,7 @@
 
 static void _take_busy_lock(rw_lock *lock)
 {
-	uint8_t prev;
+	uint32_t prev;
 	do {
 		prev = __sync_lock_test_and_set(&lock->busy, 1);
 	} while (prev);
